requirements.txt
# Telegram Bot
python-telegram-bot==22.6

# LangChain and RAG
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
chromadb>=0.5.0

# OpenAI
openai==1.59.7

# Content Processing
beautifulsoup4==4.12.3
html2text==2024.2.26
yt-dlp>=2024.1.0
pypdf==5.6.1
requests==2.32.3

# Utilities
python-dotenv==1.0.1
aiohttp==3.11.12
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
